# backend_server/openapi.swagger.yaml
openapi: 3.1.0
info:
  title: Tula-Ops API Specification
  description: |
    Backend API for Tula-Ops (Community Operations & Rural Artisan Production Management Platform).
    
    User stories covered in this specification:
    - As a Core Team Member, Volunteer, or Customer, I want to register and login to access the operations dashboard.
    - As an authenticated user, I want session management (me, logout, password change) to maintain account security.
    - As a Core Team Member, I want to manage artisan profiles (list, create, update, soft-delete) to coordinate production workflows.

    API Mapping Summary:
    - Authentication: /api/v1/auth/*
    - Artisan Management: /api/v1/artisans/*
    - Health Check: /health
  version: 1.0.0
  contact:
    name: Team-097
    email: dev@tulaindia.org
servers:
  - url: /api/v1
    description: Tula-Ops API Base Path
security:
  - BearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user & issue JWT token
      operationId: auth_login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Deactivated account access attempt
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      operationId: auth_register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error or email already exists
...
  /batches:
    post:
      tags:
        - Production Batches
      summary: Create a new production batch (Core Team only)
      operationId: batches_create
      description: |
        User story: As a Core Team Member, I want to create a new production batch linked to a farmer
        so that production from raw cotton to finished garments can be tracked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
      responses:
        '201':
          description: Batch created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      batch:
                        $ref: '#/components/schemas/ProductionBatch'
        '400':
          description: Invalid request payload

    get:
      tags:
        - Production Batches
      summary: List all production batches with optional filters
      operationId: batches_list
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [in_progress, completed, paused]
        - name: farmer_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of production batches
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      batches:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProductionBatch'
                      count:
                        type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            access_token:
              type: string
            token_type:
              type: string
              example: Bearer
            user:
              $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [core_team_member, volunteer, customer]
        phone:
          type: string
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
...
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    ProductionBatch:
      type: object
      properties:
        id:
          type: string
        farmer_id:
          type: string
        recorded_by:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [in_progress, completed, paused]
        created_at:
          type: string
    CreateBatchRequest:
      type: object
      required: [farmer_id]
      properties:
        farmer_id:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [in_progress, completed, paused]
...